requirements.txt
pyautogui==0.9.54
pillow==10.1.0
python-dotenv==1.0.0
ollama==0.1.5
requests==2.31.0
autopep8==2.0.4
numpy>=1.22.0
opencv-python>=4.6.0
easyocr>=1.7.0
paddleocr>=2.6.0
matplotlib>=3.5.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
ultralytics>=8.0.0
transformers>=4.30.0
safetensors>=0.3.1
huggingface_hub>=0.20.0 